Separate language service logic from index.ts
